fanvue-discord-sync/requirements.txt
discord.py
pyyaml
../fanvue_common
